+++ fabric-networking/src/main/resources/fabric.mod.json
@@ -1,0 +1,6 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-networking",
+  "version": "${version}",
+  "license": "Apache-2.0"
+}
