+++ fabric-renderer-api-v1/src/main/resources/fabric-renderer-api-v1.mixins.json
@@ -2,14 +2,9 @@
   "required": true,
   "package": "net.fabricmc.fabric.mixin.renderer",
   "compatibilityLevel": "JAVA_8",
-  "mixins": [
-    "MixinBlockEntity",
-    "MixinWorld"
-  ],
   "client": [
     "client.MixinBakedModel",
     "client.MixinBlockRenderManager",
-    "client.MixinChunkRendererRegion",
     "client.MixinSpriteAtlasTexture"
   ],
   "injectors": {
