+++ fabric-rendering-data-attachment-v1/build.gradle
@@ -1,0 +1,6 @@
+archivesBaseName = "fabric-rendering-data-attachment-v1"
+version = getSubprojectVersion(project, "0.1.0")
+
+dependencies {
+	compile project(path: ':fabric-api-base', configuration: 'dev')
+}
