+++ fabric-rendering-data-attachment-v1/src/main/resources/fabric.mod.json
@@ -1,0 +1,13 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-rendering-data-attachment-v1",
+  "version": "${version}",
+  "license": "Apache-2.0",
+  "depends": {
+    "fabricloader": ">=0.4.0",
+    "fabric-api-base": "*"
+  },
+  "mixins": [
+    "fabric-rendering-data-attachment-v1.mixins.json"
+  ]
+}
