+++ settings.gradle
@@ -31,6 +31,7 @@ include 'fabric-registry-sync-v0'
 include 'fabric-renderer-api-v1'
 include 'fabric-renderer-indigo'
 include 'fabric-rendering-v0'
+include 'fabric-rendering-data-attachment-v1'
 include 'fabric-rendering-fluids-v1'
 include 'fabric-resource-loader-v0'
 include 'fabric-tag-extensions-v0'
